telegram-pdf-maker-bot-2/requirements.txt
python-telegram-bot==20.4
Pillow==10.0.0
opencv-python==4.8.0
python-dotenv==1.0.0